Update se.psv
Sentio polls excluded from the polling average as they are not deemed as credible by any established Swedish media and universities. According to the University of Gothenburg, ever since Sentio has been commissioned by Nyheter Idag (since 2015), it has been impossible to get in touch with the pollster. There has been multiple evidence of faked sample sizes while in reality very small samples were used. Furthermore, according to the university Sentio polls lack any transparency and in order to get any information one has to contact the commissioner Nyheter Idag (a news site with direct links to the party SD-ECR). And as earlier said, no Swedish media but Nyheter Idag and the Sweden Democrats themselves publish, give attention to and spread Sentio polls. Polling averages by the University of Gothenburg, news network TV4 and Kantar Sifo for Sweden Radio ignore the pollster for this reason. We discussed this in the polling chat of Europeelects and there we agreed to not posting the poll anymore.